fixed collection for service multi-region tests
